% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cart_app.R
\name{cart_app}
\alias{cart_app}
\title{Shiny App Interface with rpart::rpart}
\usage{
cart_app(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Shiny app.
}
\description{
User selects response variable and predictors, app displays optimal tree in
terms of cross-validation error along with a summary table and importance
plot. Relies on \code{\link[rpart]{rpart}} for fitting and
\code{\link[rattle]{fancyRpartPlot}} for plotting.
}
