\name{ccor}
\alias{ccor}
\title{
Complete-Data Sample Correlation
}
\description{
Same as \code{\link{cor}}, but with default \code{use = "complete.obs"}.
}
\usage{
ccor(x, use = "complete.obs", ...)
}
\arguments{
  \item{x}{
See \code{\link{cor}}.
}
  \item{use}{
See \code{\link{cor}}.
}
  \item{...}{
Additional arguments to pass to \code{\link{cor}}.
}
}
\details{
NA
}
\value{
See \code{\link{cor}}.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
