\name{cquantile}
\alias{cquantile}
\title{
Complete-Data Sample Quantiles
}
\description{
Same as \code{\link{quantile}}, but with default \code{na.rm = TRUE}.
}
\usage{
cquantile(x, na.rm = TRUE, ...)
}
\arguments{
  \item{x}{
See \code{\link{quantile}}.
}
  \item{na.rm}{
See \code{\link{quantile}}.
}
  \item{...}{
Additional arguments to pass to \code{\link{quantile}}.
}
}
\details{
NA
}
\value{
See \code{\link{quantile}}.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
