\name{crange}
\alias{crange}
\title{
Complete-Data Range of Values
}
\description{
Same as \code{\link{range}}, but with default \code{na.rm = TRUE}.
}
\usage{
crange(..., na.rm = TRUE)
}
\arguments{
  \item{...}{
See \code{\link{range}}.
}
  \item{na.rm}{
See \code{\link{range}}.
}
}
\details{
NA
}
\value{
See \code{\link{range}}.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
