\name{csd}
\alias{csd}
\title{
Complete-Data Sample Standard Deviation
}
\description{
Same as \code{\link{sd}}, but with default \code{na.rm = TRUE}.
}
\usage{
csd(x, na.rm = TRUE)
}
\arguments{
  \item{x}{
See \code{\link{sd}}.
}
  \item{na.rm}{
See \code{\link{sd}}.
}
}
\details{
NA
}
\value{
See \code{\link{sd}}.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
