\name{cvar}
\alias{cvar}
\title{
Complete-Data Sample Variance
}
\description{
Same as \code{\link{var}}, but with default \code{na.rm = TRUE}.
}
\usage{
cvar(x, na.rm = TRUE, ...)
}
\arguments{
  \item{x}{
See \code{\link{var}}.
}
  \item{na.rm}{
See \code{\link{var}}.
}
  \item{...}{
Additional arguments to pass to \code{\link{var}}.
}
}
\details{
NA
}
\value{
See \code{\link{var}}.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
