\name{cweighted.mean}
\alias{cweighted.mean}
\title{
Complete-Data Weighted Arithmetic Mean
}
\description{
Same as \code{\link{weighted.mean}}, but with default \code{na.rm = TRUE}.
}
\usage{
cweighted.mean(x, w, na.rm = TRUE, ...)
}
\arguments{
  \item{x}{
See \code{\link{weighted.mean}}.
}
  \item{w}{
See \code{\link{weighted.mean}}.
}
  \item{na.rm}{
See \code{\link{weighted.mean}}.
}
  \item{...}{
Additional arguments to pass to \code{\link{weighted.mean}}.
}
}
\details{
NA
}
\value{
See \code{\link{weighted.mean}}.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
