% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvmisc-data.R
\docType{data}
\name{dat1}
\alias{dat1}
\title{Dataset with Simulated (Y, C) Values for Examples in dfa_xerrors and
logreg_xerrors}
\source{
Simulated data in R.
}
\description{
Includes 5,000 observations. The \code{Xtilde} values are stored as a
separate list called \code{\link{dat1_xtilde}}. The data were generated with
a true log-odds ratio of 0.5 for \code{X} and \code{Y}, adjusted for
\code{C}. The \code{Xtilde} measurements are subject to measurement error.
}
