% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvmisc-data.R
\docType{data}
\name{dat1_xtilde}
\alias{dat1_xtilde}
\title{Dataset with Simulated Xtilde Values for Examples in dfa_xerrors and
logreg_xerrors}
\source{
Simulated data in R.
}
\description{
Includes 5,000 observations, 30 of which have replicates. The \code{(Y, C)}
values are stored as a separate data frame called \code{\link{dat1}}. The
data were generated with a true log-odds ratio of 0.5 for \code{X} and
\code{Y}, adjusted for \code{C}. The \code{Xtilde} measurements are subject
to measurement error.
}
