% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvmisc-data.R
\docType{data}
\name{dat2_xtilde}
\alias{dat2_xtilde}
\title{Dataset with Simulated Xtilde Values for Examples in logreg_xerrors2}
\source{
Simulated data in R.
}
\description{
Includes 200 observations, 20 of which have replicates. The (Y, C) values are 
stored as a separate data frame called \code{\link{dat2}}. The data were 
generated with a true log-OR of 0.2 for X and Y, adjusted for C. The Xtilde 
measurements are subject to multiplicative lognormal measurement error.
}
