% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{Same as expand.grid but with Sequences Reversed}
\usage{
expand_grid(...)
}
\arguments{
\item{...}{Vectors you want all combinations of.}
}
\value{
Data frame.
}
\description{
Loops over the last argument, then the second-last, and so on. Should be 
faster than \code{\link[base]{expand.grid}}.
}
\examples{
# Compare expand.grid to expand_grid
expand.grid(x = c("a", "b", "c"), y = c(1, 2), z = c(TRUE, FALSE))
expand_grid(x = c("a", "b", "c"), y = c(1, 2), z = c(TRUE, FALSE))

}
