% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_facebook.R
\name{gg_facebook}
\alias{gg_facebook}
\title{Quickly Format and Save a ggplot Object for Posting to Facebook}
\usage{
gg_facebook(plot = last_plot(), filename = paste("fig1-", Sys.Date(),
  ".png", sep = ""), resize_labels = TRUE, title_size = 16,
  axis_size = 13, height = 4, width = 4/0.525, dpi = 1000)
}
\arguments{
\item{plot}{\code{\link[ggplot2]{ggplot}} object}

\item{filename}{Character string}

\item{resize_labels}{Logical value}

\item{title_size, axis_size}{Numeric value}

\item{height, width}{Numeric value}

\item{dpi}{Numeric value}
}
\description{
Optionally resizes title and axis labels and saves a 1000 dpi png 4 inches 
tall and ~7.6 inches wide via \code{\link[ggplot2]{ggsave}}. Values are 
adjustable.
}
