% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_quicksave.R
\name{gg_quicksave}
\alias{gg_quicksave}
\title{Quickly Format and Save a ggplot Object with Reasonable Defaults}
\usage{
gg_quicksave(
  filename = paste("fig1-", Sys.Date(), ".png", sep = ""),
  path = getwd(),
  plot = last_plot(),
  resize_labels = TRUE,
  title_size = 16,
  axistitle_size = 13,
  axistext_size = 9,
  height = 6,
  width = height/0.75,
  dpi = 1000
)
}
\arguments{
\item{filename}{Character string}

\item{path}{Character string}

\item{plot}{\code{\link[ggplot2]{ggplot}} object}

\item{resize_labels}{Logical value}

\item{title_size}{Numeric value}

\item{axistitle_size}{Numeric value}

\item{axistext_size}{Numeric value}

\item{height}{Numeric value}

\item{width}{Numeric value}

\item{dpi}{Numeric value}
}
\description{
Optionally resizes title and axis labels and saves a 1000 dpi png 6 inches
tall and 8 inches wide via \code{\link[ggplot2]{ggsave}}. Values are
adjustable.
}
