% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_twitter.R
\name{gg_twitter}
\alias{gg_twitter}
\title{Quickly Format and Save a ggplot Object for Posting to Twitter}
\usage{
gg_twitter(filename = paste("fig1-", Sys.Date(), ".png", sep = ""),
  path = getwd(), plot = last_plot(), resize_labels = TRUE,
  title_size = 16, axistitle_size = 13, axistext_size = 9,
  height = 5, width = height/0.5625, dpi = 600)
}
\arguments{
\item{filename}{Character string}

\item{path}{Character string}

\item{plot}{\code{\link[ggplot2]{ggplot}} object}

\item{resize_labels}{Logical value}

\item{title_size}{Numeric value}

\item{axistitle_size}{Numeric value}

\item{axistext_size}{Numeric value}

\item{height}{Numeric value}

\item{width}{Numeric value}

\item{dpi}{Numeric value}
}
\description{
Optionally resizes title and axis labels and saves a 600 dpi png 5 inches
tall and ~8.9 inches wide via \code{\link[ggplot2]{ggsave}}. Values are
adjustable.
}
