% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal.R
\name{lognormal}
\alias{lognormal}
\title{Linear Regression Model for log(Y) vs. Covariates}
\usage{
lognormal(y, x = NULL, merror = FALSE, estimate_var = TRUE, ...)
}
\arguments{
\item{y}{Numeric vector or list.}

\item{x}{Numeric vector or matrix. If \code{NULL}, model reduces to marginal
lognormal model Y ~ Lognormal(beta_0, sigsq).}

\item{merror}{Logical value for whether to model multiplicative lognormal 
measurement errors in Y.}

\item{estimate_var}{Logical value for whether to return Hessian-based
variance-covariance matrix.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List of parameter estimates, variance-covariance matrix (if
requested), AIC, and \code{\link[stats]{nlminb}} object.
}
\description{
Uses maximum likelihood to fit
Y|\strong{X} ~ Lognormal(beta_0 + \strong{beta_x}^T \strong{X}), sigsq), with 
Y precisely measured or subject to multiplicative mean-1 lognormal errors (in 
which case replicates can be incorporated by specifying \code{y} as a list).
}
