\name{movingave}
\alias{movingave}
\title{
Moving Averages
}
\description{
This function returns a vector of moving averages for input vector \code{x}.
}
\usage{
movingave(x, window)
}
\arguments{
  \item{x}{
Numeric vector.
}
  \item{window}{
Window length.
}
}
\details{
NA
}
\value{
Numeric vector of length \code{(length(x) - window + 1)} with moving averages.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National 
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}