% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_prob.R
\name{odds_prob}
\alias{odds_prob}
\title{Convert Odds to Probability}
\usage{
odds_prob(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Defined simply as \code{log(x / (x + 1))}.
}
