% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lognormal_gamma.R
\name{plot_lognormal_gamma}
\alias{plot_lognormal_gamma}
\title{Plot Fitted Lognormal and Gamma Regression Models for Y vs. X}
\usage{
plot_lognormal_gamma(x = NULL, y = NULL, n = 1000, beta_0 = 0,
  beta_x = 0.5, sigsq = 0.1, b = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}

\item{n}{Numeric value.}

\item{beta_0}{Numeric value.}

\item{sigsq}{Numeric value.}

\item{b}{Numeric value}

\item{beta_c}{Numeric value.}
}
\value{
Plot of E(X) vs. X according to each fitted model.
}
\description{
User can specify parameters under which to generate X and Y, or specify X and
Y directly (e.g. if fitting models to real data).
}
\details{
If \code{x} and \code{y} are not specified, data are generated as lognormal
or Gamma according to the following models. If \code{b} is not specified,
data are generated:

X ~ N(0, 1)
Y|X ~ Lognormal(beta_0 + beta_x X, sigsq)

If \code{b} is specified, data are generated:

X ~ N(0, 1)
Y|X ~ Gamma(exp(beta_0 + beta_x X), b)

Both models are fit using maximum likelihood, and the data are plotted with
the fitted curves for each model.
}
