\name{pooled.var}
\alias{pooled.var}
\title{
Pooled Sample Variance
}
\description{
Calculates pooled sample variance used in equal variance two-sample t-test.
}
\usage{
pooled.var(x, y, xy.integer = FALSE)
}
\arguments{
  \item{x}{
Numeric vector.
}
  \item{y}{
Numeric vector.
}
  \item{xy.integer}{
Logical value indicating whether \code{x} and \code{y} take on integer values.
}
}
\details{
NA
}
\value{
Numeric value.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
