% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_var_i.R
\name{pooled_var_i}
\alias{pooled_var_i}
\title{Pooled Sample Variance for Integer Vectors}
\usage{
pooled_var_i(x, y)
}
\arguments{
\item{x}{Integer vector.}

\item{y}{Integer vector.}
}
\value{
Numeric value.
}
\description{
Calculates pooled sample variance used in equal variance two-sample t-test. 
Runs faster than \code{\link{pooled_var_n}} when \code{x} and \code{y} are 
integer vectors, but not valid if \code{x} or \code{y} are non-integer 
vectors.
}
