% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_var_n.R
\name{pooled_var_n}
\alias{pooled_var_n}
\title{Pooled Sample Variance for Numeric Vectors}
\usage{
pooled_var_n(x, y)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Numeric value.
}
\description{
Calculates pooled sample variance used in equal variance two-sample t-test. 
For integer vectors, \code{\link{pooled_var_i}} will run faster.
}
