% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_odds.R
\name{prob_odds}
\alias{prob_odds}
\title{Convert Probability to Odds}
\usage{
prob_odds(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Defined simply as \code{x / (1 - x)}.
}
