% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ral.R
\name{ral}
\alias{ral}
\title{Remove All Labels}
\usage{
ral(x)
}
\arguments{
\item{x}{Vector or data frame.}
}
\value{
Vector or data frame.
}
\description{
Calls \code{\link[sjlabelled]{remove_all_labels}} and then restores special 
variable names.
}
