% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{range_i}
\alias{range_i}
\title{Range (Actually Minimum and Maximum) of Integer Values}
\usage{
range_i(x)
}
\arguments{
\item{x}{Integer vector or matrix.}
}
\value{
Integer vector.
}
\description{
Written in C++, this function should always run faster than 
\code{\link[base]{range}} for integer vectors/matrices. Not valid for 
non-integer objects.
}
\examples{
# In general, range_i is much faster than range
x <- rpois(1000, lambda = 5) 
all.equal(range(x), range_i(x))
benchmark(range(x), range_i(x), replications = 10000)

}
