% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_cov.R
\name{sliding_cov}
\alias{sliding_cov}
\title{Moving Covariance as Short Vector Slides Across Long Vector}
\usage{
sliding_cov(short, long)
}
\arguments{
\item{short}{Numeric vector.}

\item{long}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Uses C++ code for efficiency.
}
\examples{
short <- rnorm(4)
long <- rnorm(10)
sliding_cov(short, long)


}
