% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.R
\name{snip}
\alias{snip}
\title{Prep for Snipping or Copy/Pasting out of R}
\usage{
snip(x, ...)
}
\arguments{
\item{x}{Object to snip}

\item{...}{Arguments to pass to \code{\link[kableExtra]{kable_styling}}.}
}
\description{
Calls \code{\link[knitr]{kable}} and \code{\link[kableExtra]{kable_styling}} 
as appropriate based on class of input object and creates object ready to 
snip or copy/paste out of R. Currently works for data frames, matrices, and 
tables.
}
\examples{
# Crosstab of gears and cylinder using mtcars dataset
snip(table(mtcars$cyl, mtcars$gear))

# Piping works, but you lose the group labels
table(mtcars$cyl, mtcars$gear) \%>\% snip()


}
