% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weighted_mean_ii}
\alias{weighted_mean_ii}
\title{Weighted Arithmetic Mean for Integer Values and Integer Weights}
\usage{
weighted_mean_ii(x, w)
}
\arguments{
\item{x}{Integer vector of values.}

\item{w}{Integer vector of weights.}
}
\value{
Numeric value.
}
\description{
Written in C++, this function should always run faster than 
\code{\link{weighted.mean}}.
}
\details{
For optimal speed, choose the version of this function that matches the 
class of your \code{x} and \code{w}: 
\code{\link{weighted_mean_nn}} for numeric \code{x}, numeric \code{w} \cr 
\code{\link{weighted_mean_ni}} for numeric \code{x}, integer \code{w} \cr 
\code{\link{weighted_mean_in}} for integer \code{x}, numeric \code{w} \cr 
\code{\link{weighted_mean_ii}} for integer \code{x}, integer \code{w} \cr

These functions typically execute several times faster than the base R 
function \code{\link[stats]{weighted.mean}} and weighted average functions 
in other packages (e.g. \code{wtd.mean} in \pkg{Hmisc} and \code{wt.mean} in 
\pkg{SDMTools}).
}
\examples{
# weighted_mean_ii is typically much faster than weighted.mean
x <- rpois(1000, lambda = 5)
w <- rpois(1000, lambda = 5)
all.equal(weighted.mean(x, w), weighted_mean_ii(x, w))
benchmark(weighted.mean(x, w), weighted_mean_ii(x, w), replications = 2000)

}
