\name{dvmisc-package}
\alias{dvmisc-package}
\alias{dvmisc}
\docType{package}
\title{
Miscellaneous General-Purpose Functions
}
\description{
A collection of general-purpose R functions, most of which are simple functions
that do something convenient, e.g. easily create histogram with fitted PDF/PMF
curve, create equal-interval groups, assess linearity in logistic regression, or
extract MSE from a fitted regression model.
}
\details{
\tabular{ll}{
Package: \tab dvmisc\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2017-09-16\cr
License: \tab GPL-2\cr
}
The following functions are included:

\code{\link{list.override}}\cr
\code{\link{inside}}\cr
\code{\link{interval.groups}}\cr
\code{\link{quant.groups}}\cr
\code{\link{bmi3}}\cr
\code{\link{bmi4}}\cr
\code{\link{means.graph}}\cr
\code{\link{dots.bars}}\cr
\code{\link{logodds.graph}}\cr
\code{\link{which.min2}}\cr
\code{\link{which.max2}}\cr
\code{\link{get.mse}}\cr
\code{\link{sum}}\cr
\code{\link{cmean}}\cr
\code{\link{cweighted.mean}}\cr
\code{\link{cmedian}}\cr
\code{\link{cquantile}}\cr
\code{\link{cmin}}\cr
\code{\link{cmax}}\cr
\code{\link{crange}}\cr
\code{\link{csd}}\cr
\code{\link{cvar}}\cr
\code{\link{ccov}}\cr
\code{\link{ccor}}\cr
\code{\link{histo}}\cr
\code{\link{movingave}}
}
\author{
Dane R. Van Domelen

Maintainer: Dane R. Van Domelen <vandomed@gmail.com>
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\keyword{ package }
\examples{
NA
}
