\name{dvmisc-package}
\alias{dvmisc-package}
\alias{dvmisc}
\docType{package}
\title{
Miscellaneous General-Purpose Functions
}
\description{
A collection of general-purpose R functions, most of which are either simple functions that do something convenient (e.g. create quantile groups, assess linearity in logistic regression) or faster versions of base R functions (e.g. run length encoding).
}
\details{
\tabular{ll}{
Package: \tab dvmisc\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2017-04-29\cr
License: \tab GPL-2\cr
}
The following functions are included:

\code{\link{list.override}}\cr
\code{\link{inside}}\cr
\code{\link{interval.groups}}\cr
\code{\link{quant.groups}}\cr
\code{\link{bmi3}}\cr
\code{\link{bmi4}}\cr
\code{\link{means.graph}}\cr
\code{\link{logodds.graph}}
}
\author{
Dane R. Van Domelen

Maintainer: Dane R. Van Domelen <vandomed@gmail.com>
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\keyword{ package }
\examples{
NA
}