\name{dvmisc-package}
\alias{dvmisc-package}
\alias{dvmisc}
\docType{package}
\title{
Faster Computation of Common Statistics and Miscellaneous Functions
}
\description{
Faster versions of base R functions (e.g. mean, standard deviation, covariance,
weighted mean), mostly written in C++, along with miscellaneous functions
for various purposes (e.g. create histogram with fitted probability density
function or probability mass function curve, create body mass index groups,
assess linearity assumption in logistic regression).
}
\details{
\tabular{ll}{
Package: \tab dvmisc\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2017-11-05\cr
License: \tab GPL-2\cr
}
The following functions are included:

\code{\link{list.override}}\cr
\code{\link{inside}}\cr
\code{\link{interval.groups}}\cr
\code{\link{quant.groups}}\cr
\code{\link{bmi3}}\cr
\code{\link{bmi4}}\cr
\code{\link{means.graph}}\cr
\code{\link{dots.bars}}\cr
\code{\link{logodds.graph}}\cr
\code{\link{get.mse}}\cr
\code{\link{sum}}\cr
\code{\link{histo}}\cr
\code{\link{sum_i}}\cr
\code{\link{mean_n}}\cr
\code{\link{mean_i}}\cr
\code{\link{var_n}}\cr
\code{\link{var_i}}\cr
\code{\link{cov_n}}\cr
\code{\link{cov_i}}\cr
\code{\link{sd_n}}\cr
\code{\link{sd_i}}\cr
\code{\link{min_n}}\cr
\code{\link{min_i}}\cr
\code{\link{max_n}}\cr
\code{\link{max_i}}\cr
\code{\link{range_n}}\cr
\code{\link{range_i}}\cr
\code{\link{true_range_n}}\cr
\code{\link{true_range_i}}\cr
\code{\link{diff1_n}}\cr
\code{\link{diff1_i}}\cr
\code{\link{diff_n}}\cr
\code{\link{diff_i}}\cr
\code{\link{which_min_nv}}\cr
\code{\link{which_min_iv}}\cr
\code{\link{which_min_nm}}\cr
\code{\link{which_min_im}}\cr
\code{\link{which_max_nv}}\cr
\code{\link{which_max_iv}}\cr
\code{\link{which_max_nm}}\cr
\code{\link{which_max_im}}\cr
\code{\link{weighted_mean_nn}}\cr
\code{\link{weighted_mean_ni}}\cr
\code{\link{weighted_mean_in}}\cr
\code{\link{weighted_mean_ii}}\cr
\code{\link{pooled_var_n}}\cr
\code{\link{pooled_var_i}}\cr
\code{\link{prob_odds}}\cr
\code{\link{prob_logit}}\cr
\code{\link{odds_prob}}\cr
\code{\link{logit_prob}}
}
\author{
Dane R. Van Domelen

Maintainer: Dane R. Van Domelen <vandomed@gmail.com>
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\keyword{ package }
\examples{
NA
}
