% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvmisc.R
\docType{package}
\name{dvmisc}
\alias{dvmisc}
\alias{dvmisc-package}
\title{Convenience Functions, Moving Window Statistics, and Graphics}
\description{
Contains functions that do something convenient (e.g. create BMI categories), 
functions for calculating moving-window statistics efficiently, and functions 
for generating various figures (e.g. histograms with fitted probability 
mass/density function).
}
\details{
\tabular{ll}{
Package: \tab dvmisc \cr
Type: \tab Package \cr
Version: \tab 1.1.4 \cr
Date: \tab 2019-12-15 \cr
License: \tab GPL-3 \cr
}

See \href{https://cran.r-project.org/package=dvmisc}{CRAN documentation} for 
full list of functions.
}
\references{
Eddelbuettel, D. and Francois, R. (2011) Rcpp: Seamless R and C++ 
Integration. Journal of Statistical Software, 40(8), 1-18. 
\url{http://www.jstatsoft.org/v40/i08/}.

Eddelbuettel, D. (2013) Seamless R and C++ Integration with Rcpp. Springer, 
New York. ISBN 978-1-4614-6867-7.

Eddelbuettel, D. and Balamuta, J.J. (2017). Extending R with C++: A Brief 
Introduction to Rcpp. PeerJ Preprints 5:e3188v1. 
\url{https://doi.org/10.7287/peerj.preprints.3188v1}.

Acknowledgment: This material is based upon work supported by the 
National Science Foundation Graduate Research Fellowship under Grant No. 
DGE-0940903.
}
\author{
Dane R. Van Domelen \cr \email{vandomed@gmail.com}
}
