% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{Same as expand.grid but with Sequences Reversed}
\usage{
expand_grid(..., KEEP.OUT.ATTRS = TRUE, stringsAsFactors = TRUE)
}
\arguments{
\item{...}{See \code{\link[base]{expand.grid}}.}

\item{KEEP.OUT.ATTRS}{See \code{\link[base]{expand.grid}}.}

\item{stringsAsFactors}{See \code{\link[base]{expand.grid}}.}
}
\value{
Data frame.
}
\description{
Loops over the last argument, then the second-last, and so on. Defined simply 
as \code{rev(expand.grid(rev(list(...))))}.
}
\examples{
# Compare expand.grid to expand_grid
expand.grid(x = c("a", "b", "c"), y = c(1, 2), z = c(TRUE, FALSE))
expand_grid(x = c("a", "b", "c"), y = c(1, 2), z = c(TRUE, FALSE))

}
