% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_constantscale.R
\name{gamma_constantscale}
\alias{gamma_constantscale}
\title{Constant-Scale Gamma Model for Y vs. Covariates}
\usage{
gamma_constantscale(y, x = NULL, merror = FALSE, estimate_var = TRUE,
  integrate_tol = 1e-08, integrate_tol_hessian = integrate_tol, ...)
}
\arguments{
\item{y}{Numeric vector.}

\item{x}{Numeric vector or matrix. If \code{NULL}, model reduces to marginal
Gamma model Y ~ Gamma(exp(beta_0), b).}

\item{merror}{Logical value for whether to model multiplicative lognormal 
measurement errors in Y.}

\item{estimate_var}{Logical value for whether to return Hessian-based
variance-covariance matrix.}

\item{integrate_tol}{Numeric value specifying the \code{tol} input to
\code{\link{hcubature}}. Only used if \code{merror = TRUE}.}

\item{integrate_tol_hessian}{Same as \code{integrate_tol}, but for use when
estimating the Hessian matrix only. Sometimes more precise integration
(i.e. smaller tolerance) than used for maximizing the likelihood helps
prevent cases where the inverse Hessian is not positive definite.}

\item{...}{Additional arguments to pass to \code{\link[stats]{nlminb}}.}
}
\value{
List of parameter estimates, variance-covariance matrix (if
requested), AIC, and \code{\link[stats]{nlminb}} object.
}
\description{
Uses maximum likelihood to fit
Y|\strong{X} ~ Gamma(exp(beta_0 + \strong{beta_x}^T \strong{X}), b), with the
shape-scale (as opposed to shape-rate) parameterization described in
\code{\link[stats]{GammaDist}}. Y can be perfectly measured or subject to 
multiplicative mean-1 lognormal errors (in which case replicates can be 
incorporated by specifying \code{y} as a list).
}
