% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_facebook.R
\name{gg_facebook}
\alias{gg_facebook}
\title{Quickly Format and Save a ggplot Object for Posting to Facebook}
\usage{
gg_facebook(filename = paste("fig1-", Sys.Date(), ".png", sep = ""),
  path = getwd(), plot = last_plot(), resize_labels = TRUE,
  title_size = 16, axis_size = 13, height = 5,
  width = height/0.525, dpi = 1000)
}
\arguments{
\item{filename}{Character string}

\item{path}{Character string}

\item{resize_labels}{Logical value}

\item{title_size, axis_size}{Numeric value}

\item{height, width}{Numeric value}

\item{dpi}{Numeric value}
}
\description{
Optionally resizes title and axis labels and saves a 1000 dpi png 5 inches
tall and ~9.5 inches wide via \code{\link[ggplot2]{ggsave}}. Values are
adjustable.
}
