% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headtail.R
\name{headtail}
\alias{headtail}
\title{Return the First and Last Part of an Object}
\usage{
headtail(x, ...)
}
\arguments{
\item{x}{Input object.}

\item{...}{Additional arguments to pass to \code{\link[utils:head]{head}} and 
\code{\link[utils:head]{tail}} functions.}
}
\value{
Same class as \code{x}.
}
\description{
Simply \code{\link[utils:head]{head}} and \code{\link[utils:head]{tail}} 
combined.
}
\examples{
# Generate data from N(0, 1), sort, and look at smallest and largest 3 values
x <- rnorm(1000)
x.sorted <- sort(x)
headtail(x.sorted, 3)

}
