% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_prob.R
\name{logit_prob}
\alias{logit_prob}
\title{Convert Logit to Probability}
\usage{
logit_prob(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Defined as: \code{exp_x <- exp(x); out <- exp_x / (1 + exp_x)}. This 2-step 
approach is faster than \code{exp(x) / (1 + exp(x))} because the exponentials 
only have to be calculated once.
}
