\name{odds_prob}
\alias{odds_prob}
\title{
Convert Odds to Probability
}
\description{
Defined simply as \code{log(x / (x + 1))}.
}
\usage{
odds_prob(x)
}
\arguments{
  \item{x}{
Numeric vector.
}
}
\details{
NA
}
\value{
Numeric vector.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
