% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_var.R
\name{pooled_var}
\alias{pooled_var}
\title{Pooled Sample Variance}
\usage{
pooled_var(x, y, integer = FALSE)
}
\arguments{
\item{x, y}{Integer or numeric vectors.}

\item{integer}{Logical value for whether \code{x} and \code{y} are integer 
vectors.}
}
\value{
Numeric value.
}
\description{
Calculates pooled sample variance used in equal variance two-sample t-test.
}
