\name{pooled_var_n}
\alias{pooled_var_n}
\title{
Pooled Sample Variance for Numeric Vectors
}
\description{
Calculates pooled sample variance used in equal variance two-sample t-test. For 
integer vectors, \code{\link{pooled_var_i}} will run faster.
}
\usage{
pooled_var_n(x, y)
}
\arguments{
  \item{x}{
Numeric vector.
}
  \item{y}{
Numeric vector.
}
}
\details{
NA
}
\value{
Numeric value.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
