\name{prob_odds}
\alias{prob_odds}
\title{
Convert Probability to Odds
}
\description{
Defined simply as \code{x / (1 - x)}.
}
\usage{
prob_odds(x)
}
\arguments{
  \item{x}{
Numeric vector.
}
}
\details{
NA
}
\value{
Numeric vector.
}
\note{
NA
}
\references{
Acknowledgment: This material is based upon work supported by the National
Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\author{
Dane R. Van Domelen
}
\seealso{
NA
}
\examples{
NA
}
