% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_groups.R
\name{quant_groups}
\alias{quant_groups}
\title{Split Continuous Variable into Quantile Groups}
\usage{
quant_groups(x, groups = 4, probs = NULL, quantile.list = NULL,
  cut.list = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{groups}{Numeric value specifying number of quantile groups.}

\item{probs}{Numeric vector specifying probabilities.}

\item{quantile.list}{Arguments to pass to \code{\link[stats]{quantile}}.}

\item{cut.list}{Arguments to pass to \code{\link{cut}}.}
}
\value{
Factor variable.
}
\description{
Splits a continuous variable into quantiles groups. Basically combines 
\code{\link[stats]{quantile}} and \code{\link[base]{cut}} into a single 
function. Note that \code{\link{create_qgroups}} will likely supersede this 
function in future versions of \strong{dvmisc}.
}
\examples{
# Convert values from N(0, 1) into quintiles (i.e. 5 groups)
x <- rnorm(1000)
groups <- quant_groups(x, 5)
table(groups)

}
